chore: move liveness probes to health-check-core module
Refs: XRDDEV-3121
