+++ src/service/configuration-client/configuration-client-application/build.gradle.kts
@@ -38,7 +38,6 @@ publishing {
 }
 
 dependencies {
-  implementation(project(":lib:health-quarkus"))
   implementation(project(":lib:rpc-quarkus"))
   implementation(project(":common:common-core"))
   implementation(project(":lib:properties-quarkus"))
