+++ src/settings.gradle.kts
@@ -79,7 +79,6 @@ include("lib:bootstrap-edc-quarkus")
 include("lib:globalconf-impl")
 include("lib:globalconf-core")
 include("lib:globalconf-spring")
-include("lib:health-quarkus")
 include("lib:serverconf-impl")
 include("lib:serverconf-core")
 include("lib:serverconf-spring")
